# src/adapters/http/openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "StellarStay Hotels API",
    "version": "1.0.0",
    "description": "API documentation for StellarStay Hotels reservation system."
  },
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "Local server"
    }
  ],
  "paths": {
...
    "/api/ai/query": {
      "post": {
        "summary": "AI-powered hotel search",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/AIQueryRequest" }
            }
          }
        },
        "responses": {
          "200": {
            "description": "AI query response",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/AIQueryResponse" }
              }
            }
          },
          "400": { "description": "Invalid query format" },
          "404": { "description": "No rooms found matching criteria" },
          "500": { "description": "AI processing error" }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "AIQueryRequest": {
        "type": "object",
        "properties": {
          "query": { "type": "string" }
        },
        "required": ["query"]
      },
      "AIQueryResponse": {
        "type": "object",
        "properties": {
          "natural_language_response": { "type": "string" },
          "structured_results": { "type": "array", "items": { "type": "object" } },
          "extracted_parameters": { "type": "object" }
        }
      }
    }
  }
}